LayoutTests/imported/w3c/web-platform-tests/service-workers/service-worker/clients-get-cross-origin.https-expected.txt

CONSOLE MESSAGE: line 48: Unhandled Promise Rejection: TypeError: null is not an object (evaluating 'iframe.contentWindow.navigator.serviceWorker.controller.postMessage')
